more consistent command-line interface
Now format specifiers are mutually exclusive.  At most one of these can
be present:

    -l or --list
    -t or --table (default)
    -j or --json

Flags that control whether symbol names are shown in the output and
whether they are demangled are also mutually exclusive:

    -s or --show-symbol
    -d or --show-demangled

The default is to not show symbol names.

This commit also adds separate command-line arguments for filtering raw
and demangled symbols:

    -S <STRING> or --raw-symbol-filter <STRING>
    -D <STRING> or --demangled-symbol-filter <STRING>